refactor: split methods between two classes
