+++ async-race/src/app/components/car/car.ts
@@ -1,6 +1,6 @@
 import { CarData } from 'src/app/models/car-data'
 import { BaseComponent } from 'src/app/utils/base-component'
-import { httpFetcherEngine } from 'src/app/utils/http-fetcher-engine'
+import { httpFetcher } from 'src/app/utils/http-fetcher'
 import { bringCarBackToStart } from 'src/app/utils/bring-car-back-to-start'
 import { buildSvgSprite } from 'src/app/utils/build-svg-sprite'
 import { gameState } from 'src/app/utils/game-state'
@@ -54,7 +54,7 @@ export class Car extends BaseComponent {
 
       if (passedPath > roadLength) {
         this.isDriving = false
-        httpFetcherEngine.stopEngine(this.carData.id)
+        httpFetcher.stopEngine(this.carData.id)
       }
     }, 10)
   }
