+++ async-race/src/app/components/modify-car-input/modify-car-input.ts
@@ -1,5 +1,5 @@
 import { gameState } from 'src/app/utils/game-state'
-import { httpFetcherEngine } from 'src/app/utils/http-fetcher-engine'
+import { httpFetcher } from 'src/app/utils/http-fetcher'
 import { emitter } from 'src/app/utils/event-emitter'
 import { CarInput } from '../car-input/car-input'
 
@@ -23,7 +23,7 @@ export class ModifyCarInput extends CarInput {
     if (modifyingTarget) {
       const oldName = modifyingTarget.carData.name
 
-      await httpFetcherEngine.modifyCar(modifyingTarget.carData.id, {
+      await httpFetcher.modifyCar(modifyingTarget.carData.id, {
         name: this.nameInput.inputValue || oldName,
         color: this.colorInput.inputValue ?? '#000000',
       })
