+++ async-race/src/app/utils/http-fetcher.ts
@@ -5,10 +5,10 @@ import { CarData } from '../models/car-data'
 import { CreateCarData } from '../models/create-car-data'
 import { EngineStartedData } from '../models/engine-started-response'
 import { StartEngineReturnProps } from '../models/start-engine-return-props'
-import { WinnerData } from '../models/winner-data'
 import { gameState } from './game-state'
+import { httpFetcherWinners } from './http-fetcher-winners'
 
-class HTTPFetcherEngine {
+class HTTPFetcher {
   public async getCars({ isPaginationRequiered = true }): Promise<CarData[]> {
     const query = `?_page=${gameState.currentGaragePage}&_limit=${API.CarLimit}`
     const response = await fetch(`${API.Path}/garage${isPaginationRequiered ? query : ''}`)
@@ -63,10 +63,12 @@ class HTTPFetcherEngine {
       method: HTTPMethods.DELETE,
     })
 
-    await fetch(`${API.Path}/winners/${id}`, {
-      // перенести в фетчер для винеров
-      method: HTTPMethods.DELETE,
-    })
+    await httpFetcherWinners.deleteWinner(id)
+
+    // await fetch(`${API.Path}/winners/${id}`, {
+    //   // перенести в фетчер для винеров
+    //   method: HTTPMethods.DELETE,
+    // })
   }
 
   public async modifyCar(id: number, carData: CreateCarData): Promise<void> {
@@ -77,51 +79,6 @@ class HTTPFetcherEngine {
       body: requestBody,
     })
   }
-
-  public async getWinner(id: number): Promise<Response> {
-    const response = await fetch(`${API.Path}/winners/${id}`)
-    return response
-  }
-
-  public async getWinners(): Promise<WinnerData[]> {
-    const respone = await fetch(`${API.Path}/winners`)
-    const winnersData = respone.json()
-    return winnersData
-  }
-
-  public async setNewWinner(id: number, rideTime: number): Promise<Response> {
-    console.log('server time', rideTime)
-    const response = await fetch(`${API.Path}/winners`, {
-      method: HTTPMethods.POST,
-      headers: headersForPostMethod,
-      body: JSON.stringify({
-        id,
-        wins: 1,
-        time: rideTime,
-      }),
-    })
-
-    return response
-  }
-
-  public async updateWinnerData(id: number, rideTime: number): Promise<Response> {
-    const initialDataRewspose = await fetch(`${API.Path}/winners/${id}`)
-    const initialData: WinnerData = await initialDataRewspose.json()
-    let { wins, time } = initialData
-    wins += 1
-    time = time > rideTime ? rideTime : time
-
-    const response = await fetch(`${API.Path}/winners/${id}`, {
-      method: HTTPMethods.PUT,
-      headers: headersForPostMethod,
-      body: JSON.stringify({
-        wins,
-        time,
-      }),
-    })
-
-    return response
-  }
 }
 
-export const httpFetcherEngine = new HTTPFetcherEngine()
+export const httpFetcher = new HTTPFetcher()
