+++ async-race/src/app/utils/save-winner.ts
@@ -1,7 +1,7 @@
 import { RequestStatuses } from '../enum/request-statuses'
 import { emitter } from './event-emitter'
 import { gameState } from './game-state'
-import { httpFetcherEngine } from './http-fetcher-engine'
+import { httpFetcherWinners } from './http-fetcher-winners'
 
 export const saveWinner = async (): Promise<void> => {
   console.log(gameState.raceWinnerTime)
@@ -12,10 +12,10 @@ export const saveWinner = async (): Promise<void> => {
     return
   }
 
-  const getWinnerResponse = await httpFetcherEngine.getWinner(winnerId)
+  const getWinnerResponse = await httpFetcherWinners.getWinner(winnerId)
 
   if (getWinnerResponse.status === RequestStatuses.NotFound) {
-    const setWinnerResponse = await httpFetcherEngine.setNewWinner(winnerId, gameState.raceWinnerTime)
+    const setWinnerResponse = await httpFetcherWinners.setNewWinner(winnerId, gameState.raceWinnerTime)
     console.log(setWinnerResponse)
     emitter.emit('render winners')
     return
@@ -23,7 +23,7 @@ export const saveWinner = async (): Promise<void> => {
   console.log(getWinnerResponse)
   console.log('same winner')
 
-  const updateWinnerResponse = await httpFetcherEngine.updateWinnerData(winnerId, gameState.raceWinnerTime)
+  const updateWinnerResponse = await httpFetcherWinners.updateWinnerData(winnerId, gameState.raceWinnerTime)
   console.log(updateWinnerResponse)
 
   emitter.emit('render winners')
